# swagger.yaml
openapi: "3.0.0"
info:
  title: Learn express and swagger
  description: LCO - a course segment about writing docs
  version: 1.1.0
  contact:
    email: kacper.figa@protonmail.com
    url: "https://kacperfiga-frontend.netlify.app/"

  # servers:
  #   - url: "https://localhost:2137/api/v1"
  #     description: server for localhost - secure
  #   - url: "http://localhost:2137/api/v1"
  #     description: server for localhost - regular

servers:
  - url: "{protocol}://localhost:2137/api/{version}"
    description: for localhost
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer

paths:
  /lco:
    get:
      tags:
        - String
      summary: returns a greet message from lco
      responses:
        "200":
          description: All good and success
          content:
            application/json:
              schema:
                type: string
                example: "mystring"
        "400":
          description: Bad request
        "500":
          description: Internal server error
  /lcoobject:
    get:
      tags:
        - Objects
      summary: returns an object with course
      responses:
        "200":
          description: All fine hihi :)
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  id:
                    type: string
                  price:
                    type: number
        "400":
          description: Bad request
        "500":
          description: Internal server error
  /courses:
    get:
      summary: Get all courses
      responses:
        "200":
          description: All fine my friend
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    price:
                      type: number
          "400":
            description: Bad request
          "500":
            description: Internal server error
  /mycourses/{id}:
    get:
      summary: returns a specified course with specified id
      tags:
        - Objects
      parameters:
        - name: id
          in: path
          required: true
          default: 13
          schema:
            type: string
      responses:
        "200":
          description: returns a course object
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  id:
                    type: string
                  price:
                    type: number